fix(global): split code, docs
