+++ docs/en/architecture.md
@@ -245,7 +245,7 @@ Specifics:
 
 `analyze-waste.command.ts` orchestrates the run as a sequence of calls into two sibling modules ([ADR-0056](../adr/0056-analyze-waste-command-split.md)): `resolve-options.ts` (`resolveMinAgeDays`, `resolveExplicitScanners`, `resolveRegions`) resolves CLI options (regions, min-age, account ID) and loads the config file, and `post-analysis.ts` (`writeArtifacts`, `applyCostGate`) writes the file artifacts and applies the cost-gate threshold after the scan. The command itself delegates the actual instantiation of concrete implementations to `analyze-waste.composition.ts` through the injectable `AnalyzeDeps.createAnalysis` seam (the same seam `analyze-waste.command.spec.ts` fakes to test without AWS). Before that, it also resolves **which scanners to run**: `--all-services` or `--scanners <kinds...>` skip straight to a resolved list; otherwise, in a real terminal outside CI (and without `--silent`), an interactive `@clack/prompts` wizard (`apps/cli/src/wizard/scanner-selection.wizard.ts`, see [ADR-0041](../adr/0041-interactive-scanner-selection-wizard.md)) lets the user pick — every kind pre-checked, so Enter alone still scans everything; non-TTY/CI/`--silent` skip the wizard and run every scanner, unchanged from before this feature.
 
-`analyze-waste.composition.ts` is the only place where concrete implementations are instantiated. It's a declarative registry, not a hand-written list: `ALWAYS_ON_SCANNERS` and `LIVE_PRICING_SCANNERS` are each an array of `{ kind, create(ctx) }` entries, and `buildScanners()` is a `map`/`filter` over both (the second only when a live-pricing adapter is available), then filtered again down to the resolved scanner selection (`AnalysisContext.scannerKinds`, undefined = no filter). `assertRegistryMatchesResourceKinds()` runs at module load and throws if any `ResourceKind` is missing from, or duplicated across, the two registries — a wiring mistake fails at startup, not silently at scan time. See [ADR-0043](../adr/0043-declarative-scanner-registry.md). The `LIVE_PRICING_SCANNERS` entries (`AwsEc2UnderutilizedScanner`, `AwsRdsUnderutilizedScanner`, `AwsElastiCacheIdleScanner` and the Redshift/OpenSearch/MSK/DocumentDB/Neptune/MQ/WorkSpaces equivalents) are only built when `--live-pricing` is set: their cost estimate needs a per-instance-type/class/node-type price that the static table doesn't carry (too many distinct types to maintain), so without live pricing there is nothing reliable to report and the scanners are left out rather than registered with a zero estimate.
+`analyze-waste.composition.ts` calls into `scanner-registry.ts`, the declarative registry where concrete scanner implementations are instantiated — not a hand-written list. `ALWAYS_ON_SCANNERS` (`always-on-scanners.ts`, 30 entries) and `LIVE_PRICING_SCANNERS` (`live-pricing-scanners.ts`, 13 entries) are each an array of `{ kind, create(ctx) }` entries, split across the two files on that same always-on/live-pricing seam ([ADR-0077](../adr/0077-scanner-registry-split-on-pricing-seam.md)); `scanner-registry.ts` itself only re-exports both arrays plus the shared types and `buildScanners()`. `buildScanners()` is a `map`/`filter` over both arrays (the second only when a live-pricing adapter is available), then filtered again in `analyze-waste.composition.ts` down to the resolved scanner selection (`AnalysisContext.scannerKinds`, undefined = no filter). `assertRegistryMatchesResourceKinds()` runs at module load and throws if any `ResourceKind` is missing from, or duplicated across, the two registries — a wiring mistake fails at startup, not silently at scan time. See [ADR-0043](../adr/0043-declarative-scanner-registry.md). The `LIVE_PRICING_SCANNERS` entries (`AwsEc2UnderutilizedScanner`, `AwsRdsUnderutilizedScanner`, `AwsElastiCacheIdleScanner` and the Redshift/OpenSearch/MSK/DocumentDB/Neptune/MQ/WorkSpaces equivalents) are only built when `--live-pricing` is set: their cost estimate needs a per-instance-type/class/node-type price that the static table doesn't carry (too many distinct types to maintain), so without live pricing there is nothing reliable to report and the scanners are left out rather than registered with a zero estimate.
 
 Running `cloudrift` with no subcommand at all, in a real terminal, skips Commander entirely and hands off to `runEntryWizard()` (`apps/cli/src/wizard/entry.wizard.ts`) — a mode picker (waste / cost / trend) that gathers the same options as an equivalent flag-driven invocation and then calls `analyzeWasteCommand`/`costCommand`/`trendCommand` directly, so the wizard is purely an input-gathering layer with no duplicated business logic. Any explicit subcommand or flag, CI, or non-interactive stdout bypasses the wizard unchanged. See [ADR-0071](../adr/0071-unified-entry-wizard-bare-invocation.md).
