+++ docs/en/testing.md
@@ -54,6 +54,14 @@ Fixture provenance is recorded in each file's `source` field: 14 were captured f
 
 [`apps/cli/src/commands/analyze-waste.command.spec.ts`](../../apps/cli/src/commands/analyze-waste.command.spec.ts) drives the command with a fake `AnalyzeDeps` (no AWS), asserting format selection (table/json/markdown), exit codes (0/1/2), the `--json <file>` and `--pdf <file>` artifacts, and that a partial scan (`summary.scanErrors` non-empty) doesn't crash the command — the exit code stays driven only by the cost threshold, never by scan errors, and the incomplete-scan note (produced by the formatters, see [`waste-report.markdown-formatter.spec.ts`](../../apps/cli/src/formatters/waste-report.markdown-formatter.spec.ts)) reaches stdout. `analyze-waste.composition.ts` — the `defaultAnalyzeDeps` implementation the fake stands in for — has no spec of its own by design: it only wires real AWS adapters together (same pattern as `aws-account-id.resolver.ts`), and that wiring is exactly what the fake is meant to bypass in unit tests.
 
+### Cost analytics — `cost`/`trend`
+
+Same shape as `analyze`, one layer at a time, none of it touching AWS or real money:
+
+- **Domain/application**: [`compare-cost.use-case.spec.ts`](../../libs/cloud-cost/application/src/use-cases/compare-cost.use-case.spec.ts) covers the day-of-month window logic (current vs. previous period), clipping the previous period instead of spilling into the current month when it's shorter, propagating a `CostExplorerPort` failure unchanged, and the regression found during real-AWS verification: `changePercent` reports `null` rather than an astronomical percentage when the previous period rounds to $0.00. [`cost-trend.use-case.spec.ts`](../../libs/cloud-cost/application/src/use-cases/cost-trend.use-case.spec.ts) covers bucket-to-month mapping, service filtering, and requesting exactly `months` calendar months including the current partial one.
+- **Infra — cache**: [`cost-explorer-cache.adapter.spec.ts`](../../libs/cloud-cost/infrastructure/aws-adapter/src/cost-explorer/cost-explorer-cache.adapter.spec.ts) covers `CachedCostExplorerAdapter`'s disk cache ([ADR-0070](../adr/0070-cost-explorer-disk-cache-decorator.md)): serving repeat identical closed-range requests from disk instead of re-calling (re-billing) Cost Explorer, never caching a range touching the still-open current period, `--refresh-cache` bypassing but still refreshing the cache, per-account cache keys, and propagating a failure without caching it. `AwsCostExplorerAdapter` itself (the real `GetCostAndUsageCommand` call) has no spec of its own — same "thin real-AWS wiring, bypassed by the fake in every other test" pattern as `analyze-waste.composition.ts` above.
+- **CLI e2e**: [`cost.command.spec.ts`](../../apps/cli/src/commands/cost.command.spec.ts) and [`trend.command.spec.ts`](../../apps/cli/src/commands/trend.command.spec.ts) drive the commands with a fake `CostAnalyticsDeps` (no AWS, no Cost Explorer charge), asserting format selection (table/json), input validation (`--format`, `--fail-on-increase`, `--months`), the `--fail-on-increase`/`costIncreaseAlertPercent` gate (explicit flag overrides config; exit 2 on a spend spike; no gate when neither is set), `trend`'s cost-shorthand resolution (`ec2` → the documented Cost Explorer service name; an unresolved shorthand passes through unchanged), and `--silent` suppressing stdout entirely. Both commands call `confirmCostExplorerCharge()` directly (it's not part of the injectable `CostAnalyticsDeps` seam), but under Jest `isInteractiveTty()` is false, so it always short-circuits to "proceed" before ever reaching the `@clack/prompts` confirm call — the confirmation prompt itself has no dedicated spec and is unverified outside manual use.
+
 ## LocalStack e2e harness
 
 The specs above mock the AWS SDK, so they verify the *shape* of a query but never actually run the built CLI binary against anything. [`scripts/e2e-localstack.mjs`](../../scripts/e2e-localstack.mjs) closes that gap without real AWS cost or credentials: it starts a [LocalStack](https://www.localstack.cloud/) container (`docker-compose.localstack.yml`), seeds one wasted/optimizable resource per kind (`scripts/seed-localstack.mjs`), runs the built `cloudrift analyze` against it, asserts that every expected `kind` produced a finding, and tears the container down — even on failure. It passes `--all-services` explicitly so the run always covers every scanner regardless of the [interactive picker](../adr/0041-interactive-scanner-selection-wizard.md)'s trigger logic — belt and suspenders, since `spawnSync`'s piped stdout is never a TTY anyway.
