+++ docs/it/architettura.md
@@ -246,7 +246,7 @@ Particolarità:
 
 `analyze-waste.command.ts` orchestra l'esecuzione come una sequenza di chiamate a due moduli sorella ([ADR-0056](../adr/0056-analyze-waste-command-split.md)): `resolve-options.ts` (`resolveMinAgeDays`, `resolveExplicitScanners`, `resolveRegions`) risolve le opzioni CLI (regioni, min-age, account ID) e carica il file di config, e `post-analysis.ts` (`writeArtifacts`, `applyCostGate`) scrive gli artefatti file e applica la soglia del cost gate dopo lo scan. Il comando stesso delega l'istanziazione effettiva delle implementazioni concrete a `analyze-waste.composition.ts` tramite il seam iniettabile `AnalyzeDeps.createAnalysis` (lo stesso seam che `analyze-waste.command.spec.ts` finge per testare senza AWS). Prima di questo, risolve anche **quali scanner eseguire**: `--all-services` o `--scanners <kinds...>` saltano direttamente a un elenco risolto; altrimenti, in un vero terminale fuori da CI (e senza `--silent`), un wizard interattivo `@clack/prompts` (`apps/cli/src/wizard/scanner-selection.wizard.ts`, vedi [ADR-0041](../adr/0041-interactive-scanner-selection-wizard.md)) lascia scegliere all'utente — ogni kind pre-selezionato, così anche solo Invio scansiona comunque tutto; non-TTY/CI/`--silent` saltano il wizard ed eseguono ogni scanner, invariato rispetto a prima di questa funzionalità.
 
-`analyze-waste.composition.ts` è l'unico punto in cui le implementazioni concrete vengono istanziate. È un registry dichiarativo, non una lista scritta a mano: `ALWAYS_ON_SCANNERS` e `LIVE_PRICING_SCANNERS` sono ciascuno un array di entry `{ kind, create(ctx) }`, e `buildScanners()` è un `map`/`filter` su entrambi (il secondo solo se è disponibile un adapter di live pricing), filtrato poi ancora secondo la selezione scanner risolta (`AnalysisContext.scannerKinds`, undefined = nessun filtro). `assertRegistryMatchesResourceKinds()` gira al module load e lancia un errore se un `ResourceKind` manca da entrambi i registry, o è duplicato tra i due — un errore di wiring fallisce all'avvio, non silenziosamente durante la scansione. Vedi [ADR-0043](../adr/0043-declarative-scanner-registry.md). Le entry di `LIVE_PRICING_SCANNERS` (`AwsEc2UnderutilizedScanner`, `AwsRdsUnderutilizedScanner`, `AwsElastiCacheIdleScanner` e gli equivalenti Redshift/OpenSearch/MSK/DocumentDB/Neptune/MQ/WorkSpaces) vengono costruite solo se `--live-pricing` è attivo: la loro stima di costo richiede un prezzo per instance type/classe/node type che il listino statico non contiene (troppi tipi distinti da mantenere), quindi senza prezzi live non c'è nulla di affidabile da riportare e gli scanner vengono esclusi piuttosto che registrati con una stima a zero.
+`analyze-waste.composition.ts` chiama `scanner-registry.ts`, il registry dichiarativo dove le implementazioni concrete degli scanner vengono istanziate — non una lista scritta a mano. `ALWAYS_ON_SCANNERS` (`always-on-scanners.ts`, 30 entry) e `LIVE_PRICING_SCANNERS` (`live-pricing-scanners.ts`, 13 entry) sono ciascuno un array di entry `{ kind, create(ctx) }`, divisi tra i due file proprio su quel seam always-on/live-pricing ([ADR-0077](../adr/0077-scanner-registry-split-on-pricing-seam.md)); `scanner-registry.ts` stesso si limita a ri-esportare i due array più i tipi condivisi e `buildScanners()`. `buildScanners()` è un `map`/`filter` su entrambi gli array (il secondo solo se è disponibile un adapter di live pricing), filtrato poi ancora in `analyze-waste.composition.ts` secondo la selezione scanner risolta (`AnalysisContext.scannerKinds`, undefined = nessun filtro). `assertRegistryMatchesResourceKinds()` gira al module load e lancia un errore se un `ResourceKind` manca da entrambi i registry, o è duplicato tra i due — un errore di wiring fallisce all'avvio, non silenziosamente durante la scansione. Vedi [ADR-0043](../adr/0043-declarative-scanner-registry.md). Le entry di `LIVE_PRICING_SCANNERS` (`AwsEc2UnderutilizedScanner`, `AwsRdsUnderutilizedScanner`, `AwsElastiCacheIdleScanner` e gli equivalenti Redshift/OpenSearch/MSK/DocumentDB/Neptune/MQ/WorkSpaces) vengono costruite solo se `--live-pricing` è attivo: la loro stima di costo richiede un prezzo per instance type/classe/node type che il listino statico non contiene (troppi tipi distinti da mantenere), quindi senza prezzi live non c'è nulla di affidabile da riportare e gli scanner vengono esclusi piuttosto che registrati con una stima a zero.
 
 Lanciare `cloudrift` senza alcun sottocomando, in un vero terminale, salta del tutto Commander e passa la mano a `runEntryWizard()` (`apps/cli/src/wizard/entry.wizard.ts`) — un mode picker (waste / cost / trend) che raccoglie le stesse opzioni di un'invocazione equivalente guidata da flag e poi chiama direttamente `analyzeWasteCommand`/`costCommand`/`trendCommand`, quindi il wizard è puramente uno strato di raccolta input, senza logica di business duplicata. Qualunque sottocomando esplicito o flag, CI, o stdout non interattivo aggirano il wizard senza alcuna modifica. Vedi [ADR-0071](../adr/0071-unified-entry-wizard-bare-invocation.md).
