+++ docs/it/test.md
@@ -54,6 +54,14 @@ La provenienza di ogni fixture è registrata nel suo campo `source`: 14 sono sta
 
 [`apps/cli/src/commands/analyze-waste.command.spec.ts`](../../apps/cli/src/commands/analyze-waste.command.spec.ts) pilota il comando con un `AnalyzeDeps` fake (niente AWS), verificando la scelta del formato (table/json/markdown), gli exit code (0/1/2), gli artefatti `--json <file>` e `--pdf <file>`, e che uno scan parziale (`summary.scanErrors` non vuoto) non faccia crashare il comando — l'exit code resta dettato solo dalla soglia di costo, mai dagli errori di scan, e la nota di scan incompleto (prodotta dai formatter, vedi [`waste-report.markdown-formatter.spec.ts`](../../apps/cli/src/formatters/waste-report.markdown-formatter.spec.ts)) arriva fino a stdout. `analyze-waste.composition.ts` — l'implementazione `defaultAnalyzeDeps` che il fake sostituisce — non ha un suo spec per scelta: si limita a cablare insieme gli adapter AWS reali (stesso pattern di `aws-account-id.resolver.ts`), ed è esattamente quel cablaggio che il fake serve a bypassare nei test unitari.
 
+### Cost analytics — `cost`/`trend`
+
+Stessa forma di `analyze`, un layer alla volta, nessuno dei quali tocca AWS o soldi veri:
+
+- **Domain/application**: [`compare-cost.use-case.spec.ts`](../../libs/cloud-cost/application/src/use-cases/compare-cost.use-case.spec.ts) copre la logica della finestra giorno-per-giorno (periodo corrente vs. precedente), il clipping del periodo precedente invece di sconfinare nel mese corrente quando è più corto, la propagazione invariata di un fallimento di `CostExplorerPort`, e la regressione trovata durante la verifica su AWS reale: `changePercent` riporta `null` invece di una percentuale astronomica quando il periodo precedente arrotonda a $0.00. [`cost-trend.use-case.spec.ts`](../../libs/cloud-cost/application/src/use-cases/cost-trend.use-case.spec.ts) copre la mappatura bucket→mese, il filtro per servizio, e la richiesta di esattamente `months` mesi solari incluso quello corrente parziale.
+- **Infra — cache**: [`cost-explorer-cache.adapter.spec.ts`](../../libs/cloud-cost/infrastructure/aws-adapter/src/cost-explorer/cost-explorer-cache.adapter.spec.ts) copre la cache su disco di `CachedCostExplorerAdapter` ([ADR-0070](../adr/0070-cost-explorer-disk-cache-decorator.md)): servire richieste identiche ripetute su un range chiuso dalla cache invece di richiamare (e rifatturare) Cost Explorer, non cachare mai un range che tocca il periodo corrente ancora aperto, `--refresh-cache` che bypassa ma comunque aggiorna la cache, chiavi di cache per account, e la propagazione di un fallimento senza cacharlo. `AwsCostExplorerAdapter` stesso (la vera chiamata `GetCostAndUsageCommand`) non ha un suo spec — stesso pattern "wiring AWS reale sottile, bypassato dal fake in ogni altro test" di `analyze-waste.composition.ts` sopra.
+- **CLI e2e**: [`cost.command.spec.ts`](../../apps/cli/src/commands/cost.command.spec.ts) e [`trend.command.spec.ts`](../../apps/cli/src/commands/trend.command.spec.ts) pilotano i comandi con un `CostAnalyticsDeps` fake (niente AWS, nessun addebito Cost Explorer), verificando la scelta del formato (table/json), la validazione dell'input (`--format`, `--fail-on-increase`, `--months`), il gate `--fail-on-increase`/`costIncreaseAlertPercent` (il flag esplicito sovrascrive la config; exit 2 su un picco di spesa; nessun gate se nessuno dei due è impostato), la risoluzione dello shorthand di `trend` (`ec2` → il nome di servizio Cost Explorer documentato; uno shorthand non risolto passa invariato), e `--silent` che sopprime interamente lo stdout. Entrambi i comandi chiamano `confirmCostExplorerCharge()` direttamente (non fa parte del seam iniettabile `CostAnalyticsDeps`), ma sotto Jest `isInteractiveTty()` è false, quindi va sempre in cortocircuito su "procedi" prima di raggiungere la chiamata `confirm` di `@clack/prompts` — il prompt di conferma in sé non ha uno spec dedicato e non è verificato fuori dall'uso manuale.
+
 ## Harness e2e su LocalStack
 
 Gli spec sopra mockano l'SDK AWS, quindi verificano la *forma* di una query ma non lanciano mai davvero il binario CLI buildato contro qualcosa. [`scripts/e2e-localstack.mjs`](../../scripts/e2e-localstack.mjs) chiude questo gap senza costi né credenziali AWS reali: avvia un container [LocalStack](https://www.localstack.cloud/) (`docker-compose.localstack.yml`), semina una risorsa sprecata/ottimizzabile per ogni kind (`scripts/seed-localstack.mjs`), lancia `cloudrift analyze` buildato contro quel container, verifica che ogni kind atteso produca un finding, e smonta il container — anche in caso di fallimento. Passa `--all-services` esplicitamente così l'esecuzione copre sempre tutti gli scanner indipendentemente dalla logica di trigger del [picker interattivo](../adr/0041-interactive-scanner-selection-wizard.md) — cintura e bretelle, dato che lo stdout in pipe di `spawnSync` non è comunque mai un TTY.
